web_search/pyproject.toml
[project]
name = "web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.5.0",
    "linkup-sdk>=0.2.8",
    "openai-agents>=0.2.11",
]
